+++ spec/rubocop/cop/performance/delete_suffix_spec.rb
@@ -15,6 +15,18 @@ RSpec.describe RuboCop::Cop::Performance::DeleteSuffix, :config do
         str.gsub!(/suffix\\z/, '')
       RUBY
     end
+
+    it "does not register an offense when using `sub(/suffix\z/, '')`" do
+      expect_no_offenses(<<~RUBY)
+        str.sub(/suffix\\z/, '')
+      RUBY
+    end
+
+    it "does not register an offense when using `sub!(/suffix\z/, '')`" do
+      expect_no_offenses(<<~RUBY)
+        str.sub!(/suffix\\z/, '')
+      RUBY
+    end
   end
 
   context 'TargetRubyVersion >= 2.5', :ruby25 do
@@ -34,6 +46,28 @@ RSpec.describe RuboCop::Cop::Performance::DeleteSuffix, :config do
         expect_offense(<<~RUBY)
           str.gsub!(/suffix\\z/, '')
               ^^^^^ Use `delete_suffix!` instead of `gsub!`.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          str.delete_suffix!('suffix')
+        RUBY
+      end
+
+      it "registers an offense and corrects when `sub(/suffix\z/, '')`" do
+        expect_offense(<<~RUBY)
+          str.sub(/suffix\\z/, '')
+              ^^^ Use `delete_suffix` instead of `sub`.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          str.delete_suffix('suffix')
+        RUBY
+      end
+
+      it "registers an offense and corrects when `sub!(/suffix\z/, '')`" do
+        expect_offense(<<~RUBY)
+          str.sub!(/suffix\\z/, '')
+              ^^^^ Use `delete_suffix!` instead of `sub!`.
         RUBY
 
         expect_correction(<<~RUBY)
@@ -64,6 +98,28 @@ RSpec.describe RuboCop::Cop::Performance::DeleteSuffix, :config do
           str.delete_suffix!('suffix')
         RUBY
       end
+
+      it 'registers an offense and corrects when using `sub`' do
+        expect_offense(<<~RUBY)
+          str.sub(/suffix$/, '')
+              ^^^ Use `delete_suffix` instead of `sub`.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          str.delete_suffix('suffix')
+        RUBY
+      end
+
+      it 'registers an offense and corrects when using `sub!`' do
+        expect_offense(<<~RUBY)
+          str.sub!(/suffix$/, '')
+              ^^^^ Use `delete_suffix!` instead of `sub!`.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          str.delete_suffix!('suffix')
+        RUBY
+      end
     end
 
     context 'when using non-ending pattern' do
@@ -76,6 +132,18 @@ RSpec.describe RuboCop::Cop::Performance::DeleteSuffix, :config do
       it 'does not register an offense when using `gsub!`' do
         expect_no_offenses(<<~RUBY)
           str.gsub!(/pattern/, '')
+        RUBY
+      end
+
+      it 'does not register an offense when using `sub`' do
+        expect_no_offenses(<<~RUBY)
+          str.sub(/pattern/, '')
+        RUBY
+      end
+
+      it 'does not register an offense when using `sub!`' do
+        expect_no_offenses(<<~RUBY)
+          str.sub!(/pattern/, '')
         RUBY
       end
     end
@@ -92,6 +160,18 @@ RSpec.describe RuboCop::Cop::Performance::DeleteSuffix, :config do
           str.gsub!(/\\Asuffix\\z/, '')
         RUBY
       end
+
+      it 'does not register an offense and corrects when using `sub`' do
+        expect_no_offenses(<<~RUBY)
+          str.sub(/\\Asuffix\\z/, '')
+        RUBY
+      end
+
+      it 'does not register an offense and corrects when using `sub!`' do
+        expect_no_offenses(<<~RUBY)
+          str.sub!(/\\Asuffix\\z/, '')
+        RUBY
+      end
     end
 
     context 'when using a non-blank string as replacement string' do
@@ -104,6 +184,18 @@ RSpec.describe RuboCop::Cop::Performance::DeleteSuffix, :config do
       it 'does not register an offense and corrects when using `gsub!`' do
         expect_no_offenses(<<~RUBY)
           str.gsub!(/suffix\\z/, 'foo')
+        RUBY
+      end
+
+      it 'does not register an offense and corrects when using `sub`' do
+        expect_no_offenses(<<~RUBY)
+          str.sub(/suffix\\z/, 'foo')
+        RUBY
+      end
+
+      it 'does not register an offense and corrects when using `sub!`' do
+        expect_no_offenses(<<~RUBY)
+          str.sub!(/suffix\\z/, 'foo')
         RUBY
       end
     end